Rename ContactServiceInfo and make more reusable
